requirements.txt
accelerate==0.30.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diffusers==0.28.2
filelock==3.14.0
Flask==3.0.3
fsspec==2024.6.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psutil==5.9.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sympy==1.12.1
tbb==2021.12.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
